change money properties to decimal; add order total; update invoice report;

+++ Order.Management/Shapes/Circle.cs
@@ -4,7 +4,7 @@
     {
         public override ShapeName Name => ShapeName.Circle;
 
-        public override int Price => 3;
+        public override decimal Price => 3;
 
         public Circle(int numberOfRedCircles, int numberOfBlueCircles, int numberOfYellowCircles) : base(numberOfRedCircles, numberOfBlueCircles, numberOfYellowCircles) { }
     }
