+++ Order.Management/Shapes/Square.cs
@@ -4,7 +4,7 @@
     {
         public override ShapeName Name => ShapeName.Square;
 
-        public override int Price => 1;
+        public override decimal Price => 1;
 
         public Square(int numberOfRedSquares, int numberOfBlueSquares, int numberOfYellowSquares) : base(numberOfRedSquares, numberOfBlueSquares, numberOfYellowSquares) { }
     }
